Use `BoardCell` component from `board_row` function
Eliminating the need for the `board_cell` function by moving some
filtering logic for `CellUpdate::All` into `BoardCell::view`.
