+++ ui-rs/crates/mines_mogwai/src/components/game.rs
@@ -1,39 +1,6 @@
 use crate::model::{CellInteract, CellUpdate};
 use mogwai::prelude::*;
 
-/// Create a `<td>` representing a single game cell. `coords` are expected to be `(column, row)` or
-/// `(x, y)` in the grid. Interactions are transmitted to `tx` and new text to display is received
-/// by `rx`.
-#[allow(unused_braces)]
-fn board_cell<'a>(
-    coords: (usize, usize, String),
-    tx: &Transmitter<CellInteract>,
-    rx: &Receiver<CellUpdate>,
-) -> ViewBuilder<HtmlElement> {
-    use crate::components::cell::{BoardCell, BoardValue};
-    let (col, row, initial_value) = coords;
-    let tx_in = Transmitter::new();
-    let rx_value: Receiver<BoardValue> = rx.branch_filter_map(move |update| match update {
-        CellUpdate::All { cells } => cells
-            .get(row)
-            .map(|r| r.get(col))
-            .flatten()
-            .map(|s| s.into()),
-        CellUpdate::Single {
-            row: y,
-            column: x,
-            value,
-        } if *x == col && *y == row => Some(value.into()),
-        _ => None,
-    });
-    let c = Gizmo::from_parts(
-        BoardCell::new((col, row), &initial_value, tx.clone()),
-        tx_in,
-        rx_value,
-    );
-    c.view_builder()
-}
-
 /// Create a `<tr>` representing a row of game cells. Interactions are transmitted to `tx` and new
 /// text to display is received by `rx`.
 fn board_row<'a>(
@@ -42,10 +9,15 @@ fn board_row<'a>(
     tx: &Transmitter<CellInteract>,
     rx: &Receiver<CellUpdate>,
 ) -> ViewBuilder<HtmlElement> {
-    let children = initial_cells
-        .into_iter()
-        .enumerate()
-        .map(|(col, value)| board_cell((col, row, value), tx, rx));
+    use crate::components::cell::BoardCell;
+    let children = initial_cells.into_iter().enumerate().map(|(col, value)| {
+        Gizmo::from_parts(
+            BoardCell::new((col, row), &value, tx.clone()),
+            Transmitter::new(),
+            rx.branch(),
+        )
+        .view_builder()
+    });
     let mut tr = builder! { <tr /> };
     for child in children {
         tr.with(child);
